refactored(api): 使用 router 重构 example 相关 api
